Разнес создание объектов и интерфейса в разные классы. Исправил баг с анимацией у NPC
