+++ core/src/com/nuclearthinking/myheroagency/model/entity/systems/NpcSystem.java
@@ -2,13 +2,9 @@ package com.nuclearthinking.myheroagency.model.entity.systems;
 
 import com.badlogic.ashley.core.Entity;
 import com.badlogic.ashley.core.Family;
-import com.nuclearthinking.myheroagency.model.entity.AnimationState;
-import com.nuclearthinking.myheroagency.model.entity.Components;
 import com.nuclearthinking.myheroagency.model.entity.components.BodyComponent;
 import com.nuclearthinking.myheroagency.model.entity.components.NpcComponent;
 import com.nuclearthinking.myheroagency.model.entity.components.StateComponent;
-import lombok.NonNull;
-import lombok.val;
 
 /**
  * Created by mkuksin on 26.09.2016.
@@ -24,9 +20,7 @@ public class NpcSystem extends ActorSystem {
 
     @Override
     protected void processEntity(Entity entity, float deltaTime) {
-        @NonNull val state = Components.STATE.get(entity);
 
-        state.set(AnimationState.IDLE.getValue());
     }
 
     public void dialog(){
